# src/main/resources/swagger/openklips-server-api.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: OpenKlips-server
  license:
    name: GNU General Public License 3
    url: https://www.gnu.org/licenses/gpl-3.0.html
paths:
  /user/{username}:
    get:
      summary: Get the user for a given username
      operationId: getUser
      tags:
        - user
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: A user object
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDto"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
components:
  schemas:
    UserDto:
      required:
        - id
        - username
        - firstName
        - lastName
        - address
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        title:
          type: string
        address:
          $ref: '#/components/schemas/AddressDto'
        roles:
          type: array
          items:
            $ref: '#/components/schemas/RoleDto'

    AddressDto:
      properties:
        id:
          type: integer
          format: int64
        country:
          type: string
        zipCode:
          type: string
        city:
          type: string
        streetName:
          type: string
        houseNumber:
          type: string
        additionalDetails:
          type: string

    RoleDto:
      properties:
        id:
          type: integer
          format: int64
        roleType:
          $ref: '#/components/schemas/RoleType'

    RoleType:
      type: string
      enum:
        - student
        - instructor

    Error:
      required:
        - message
      properties:
        message:
          type: string